AqXCo 0.54
Modificado:
-facadeMotorGrafico
-diccionarioEnCo
-entityManager
-camaraSystem
-moveSystem
-playerSystem
-system
-main

+++ DESARROLLO/HITO1/ITERACION 3/ArquitecturaXComponentes/entitySystem/systems/camaraSystem.h
@@ -23,7 +23,7 @@ public:
     camaraSystem(const camaraSystem& orig);
     camaraSystem(entityManager*);
     virtual ~camaraSystem();
-    void update(double, unsigned int);
+    void update(unsigned int, unsigned int);
     void actualizarCamara(unsigned int, char, double);
 private:
